https://www.youtube.com/playlist?list=PLC3y8-rFHvwiWPS2RO3BKotLRfgg_8WEo React TypeScript Webpack Setup
I haven't gotten it to work yet, but this course is really good, and it shouldn't be too hard to get what I want through it.

But getting this project to run with Webpack is besides the point.

What I want is to convert it to Angular.

https://youtu.be/Lzh8A0p3z8g
WebPack Tutorial | WebPack Tutorial for Beginners | WebPack in Angular

Now that I've studied Webpack for a while and understand the basics of it, I feel more confident about watching this older course.

Let's see whether it gives me any insight into it.